Change some of the Int values to Double
Some of the values were coming as Double from the API so I changed those. Although most of the WeatherModels should change accordingly. Currently I'm just trying to make the UI work with the sample data. But a clean-up at the Models needed. I know my fault but I'm just excited to see results 🤩

+++ Weather App/Models/WeatherModel/Current/Current.swift
@@ -19,7 +19,7 @@ struct Current: Codable {
     let windDir: WindDir?
     let pressureMB: Int?
     let pressureIn: Double?
-    let precipMm, precipIn, humidity, cloud: Int?
+    let precipMm, precipIn, humidity, cloud: Double?
     let feelslikeC: Int?
     let feelslikeF: Double?
     let visKM, visMiles, uv: Int?
